CAMBIOS REALIZADOS

+++ src/main/java/dao/MedicoDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicoDao extends JpaRepository<Medico, Integer> {
+public interface MedicoDao extends JpaRepository<Medico, Long> {
 }
